Изменена цена (Float -> Int; $ -> Руб.), Исправлена кнопка "Записатьс…

+++ medic/models.py
@@ -8,7 +8,7 @@ class Procedure(models.Model):
     name = models.CharField(max_length=150, verbose_name='наименование')
     description = models.TextField(verbose_name='описание', **NULLABLE)
     picture = models.ImageField(upload_to='procedure/', verbose_name='изображение', **NULLABLE)
-    price = models.FloatField(verbose_name='цена')
+    price = models.IntegerField(verbose_name='цена')
 
     def __str__(self):
         return f'{self.name}'
